Data retrieval working for Postgres

+++ backend/spring/pulse/src/main/java/dev/thilanka/netrics/dto/KpiSnapshotDto.java
@@ -6,6 +6,6 @@ public record KpiSnapshotDto(
          Double value,
          Double previousValue,
          Double difference,
-         Long improved
+         Integer improved
 ) {
 }
